+++ src/app/shared/beancount-file-picker.ts
@@ -1,4 +1,5 @@
 import { ImageAsset } from 'tns-core-modules/image-asset';
+import { alert } from 'tns-core-modules/ui/dialogs';
 import { ImagePicker, ImagePickerMediaType } from 'nativescript-imagepicker';
 
 class FilePicker extends ImagePicker {
@@ -18,6 +19,12 @@ export function openFilePicker(): Promise<string> {
     });
     return filePicker.authorize()
         .then(() => filePicker.present())
-        .then((selection: ImageAsset[]) => selection[0].android)
+        .then((selection: ImageAsset[]) => {
+            const filePath = selection[0].android;
+            if (!filePath) {
+                alert('Failed to access the selected file.');
+            }
+            return filePath;
+        })
         .catch((error) => console.warn(error));
 }